doc: Make further improvements for Konsole etc
Try to make things a little clearer with respect to support for the
different kinds of terminals. I don't want to entirely purge the docs of
all references to iTerm (because that's still my principal test
environment), but I do want it to be clear that you don't have to use
iTerm.

Note that there is one place here where I am simplifying things in the
name of readability, while sacrificing a little bit of precision:
namely, I talk about cursor shape values getting passed through as-is to
iTerm and Konsole, and don't mention the re-mapping that occurs for the
benefit of VTE terminals. If this ever ends up confusing anybody, I'll
amend it.