Guard against "Unknown option: t_SR" for older Vim
After updating from 928b3bfb to 10564cb4 I started getting the following
error when starting Vim:

> Error detected while processing
> /path/to/.vim/bundle/terminus/plugin/terminus.vim:
> line   58:
> E355: Unknown option: t_SR

It appears that the t_SR reference was added by 640509f8.

I was running an old version of Vim for some reason (7.4.488) and
upgrading (to 7.4.873) made this message disappear. It looks like `t_SR`
was added in patch 687 [0], so we can be more hospitable to people who
aren't on the latest by adding a guard for this patch.

[0]: https://github.com/vim/vim/releases/tag/v7.4.687