Accept tmux TERM as synonym for screen
Just moved to the `master` branch (pre-2.1) of the tmux repo, and saw
that it is now recommending[1] people set up a `tmux` terminfo entry if
they want to enjoy italics support (and of course, I do).

So, this commit replaces the `s:screen` variable with `s:screenish`, and
it will be true for either screen or tmux.

[1]: https://github.com/tmux/tmux/blob/f69e09a67eaecec96e53705d45c18a5f0639e308/FAQ#L355-L383