Eliminate annoying delay on leaving insert mode
Vim would wait here (`'timeoutlen'`, usually 1 second) to see if the
`<Esc>` was going to end up being `<Esc>[something]`.

According to this page:

  https://www.johnhawthorn.com/2012/09/vi-escape-delays/

You might also need to stop Zsh from imposing a delay with:

  KEYTIMEOUT=1

And ditto for tmux:

  set -s escape-time 0

But neither of those was necessary for me.

So, fix the delay by setting up a function-key binding, and while we're
at it, reduce the `'ttimeoutlen'` from 50 to 10ms, as it's 2016 and
"10ms should be enough for anybody" (ie. almost an entire frame of
16.67ms at 60fps).