Work around missing/late FocusGained/FocusLost events
I've noticed that FocusGained/FocusLost events can be slow to arrive,
and sometimes may not arrive at all, when they are the result of
creating or destroying a new tmux pane.

Unsure of why this might be -- it could be a bug in tmux, Vim, iTerm, or
my own config -- I asked on the tmux issue tracker:

https://github.com/tmux/tmux/issues/577

As noted there, you can easily verify that tmux is always sending the
right focus events by doing this in a pane:

  printf '\033[?1004h'; cat

(ie. enable focus reporting in that pane and then have `cat` echo back
whatever events it gets back on the input stream). When you do this, you
can clearly see the right sequences coming back (`Esc[O` and `Esc[I`) as
soon as you create a split (shifting focus to it) or destroy it
(shifting focus back).

So I then started looking into Vim itself. I tried various tweaks of
`'timeout'`, `'timeoutlen'`, `'ttimeout'`, `'ttimeoutlen'` and co. I
also checked out `'updatetime'`, `'nolazyredraw'`, `'esckeys'` and
various other suspects. Additionally, I tried binding directly to
`Esc[O` rather than those virtual function keys, reordering the
autocommands to make the Focus ones run first, and disabling other
autocommands that might be interfering.

I added logging to Vim itself and recompiled it, and found that none of
the focus-related code is invoked in the terminal, which is why the
`Esc[O` stuff needs to be bound. I tried two other plug-ins that do
focus managment (Vitality[0] and vim-tmux-focus-events[1]) and found
that they exhibited the same behavior. I even tried Neovim, which has
built in focus support, and saw that it was also affected; additionally,
I looked at the implementation[2] and saw that it was implemented by
peeking ahead in the input stream, which would suggest that even if a
similar implementation were added to Vim, then we'd still have the same
problem because something else is at play.

Ruled out iTerm by doing a `touch ~/.vim/.assume-iterm` and running in
Terminal.app. Focus mostly works there, but notably, it is not immune to
this issue either.

Confirmed that I could fake focus events by doing something like this:

  tmux send-keys -t dot:1.1 Escape '[I'

But that didn't really help me much as my growing suspicion was that
something about the split creation/destruction was leading to a
VimResized event which was somehow interrupting the detection and
processing of the focus key sequence (although changing the timeouts and
even disabling timeouts entirely had no effect).

So all this leads me to this hack: we subscribe to VimResized events,
and when we get one, we check if we have focus; depending on whether we
do or don't we fire the corresponding FocusGained/FocusLost autocommand.
If we can't determine that for whatever reason, we do nothing.

The problem here is we slow down resizes. I would have liked to throttle
the events with `CursorHold`, but that doesn't really work because if
the user is idle those events get paused until the next key press. So, I
am going to commit this "eager" version, and then add a better,
throttled version in the next commit that uses timers (available in Vim
version 8, so will need feature detection).

[0]: https://github.com/sjl/vitality.vim
[1]: https://github.com/tmux-plugins/vim-tmux-focus-events
[2]: https://github.com/neovim/neovim/pull/3488