Implement throttling using timers
This ends up sucking more than you would think. I set this up with a
500ms time interval, and observed that actual time elapsed was typically
several seconds and was as high as 9 seconds. According to the help,
this is expected:

    {time} is the waiting time in milliseconds. This is the
    minimum time before invoking the callback.  When the system is
    busy or Vim is not waiting for input the time will be longer.

So, might end up adding a setting for this and letting the user choose
between responsive resizing and responsive focus/blur.